requirements.txt
google-api-python-client
google-genai>=0.3.0
streamlit
Pillow
google-auth
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit-js-eval
requests
